+++ activejob/lib/active_job/base.rb
@@ -8,6 +8,7 @@ require "active_job/enqueuing"
 require "active_job/execution"
 require "active_job/callbacks"
 require "active_job/exceptions"
+require "active_job/debouncing"
 require "active_job/log_subscriber"
 require "active_job/structured_event_subscriber"
 require "active_job/logging"
@@ -69,6 +70,7 @@ module ActiveJob # :nodoc:
     include Execution
     include Callbacks
     include Exceptions
+    include Debouncing
     include Instrumentation
     include Logging
     include ExecutionState